# apps/Gec_Server_C/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GECChain MCP API (Aptos Move)",
    "version": "1.0.0",
    "description": "MVP OpenAPI spec aligned with the current Move modules: gec_certificate, gec_marketplace, gec_audit. Intended for MCP tool integration (agent -> API -> Aptos tx)."
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local backend API"
    },
    {
      "url": "https://YOUR_DOMAIN",
      "description": "Deployed backend API"
    }
  ],
...
  "paths": {
    "/certificates/init": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_init",
        "summary": "Initialize CertRegistry under sender address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitRegistryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/certificates/add-issuer": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_add_issuer",
        "summary": "Whitelist an issuer address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddIssuerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/certificates/remove-issuer": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_remove_issuer",
        "summary": "Remove an issuer from whitelist",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RemoveIssuerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/certificates/create": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_create",
        "summary": "Create (issue) a certificate (issuer-only on-chain)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/certificates/transfer": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_transfer",
        "summary": "Transfer a certificate to a recipient",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransferCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/certificates/claim": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_claim",
        "summary": "Claim/retire a certificate (consume for reporting)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/certificates/cancel": {
      "post": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_cancel",
        "summary": "Cancel/void a certificate (invalidate)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/chain/certificates/{cert_identifier}": {
      "get": {
        "tags": [
          "Certificates"
        ],
        "operationId": "cert_view",
        "summary": "Fetch an on-chain certificate record via view function (full display fields)",
        "parameters": [
          {
            "name": "cert_identifier",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Certificate identifier: numeric id (e.g. 123) or display id (e.g. GEC-2026-000123)"
          },
          {
            "name": "registry_addr",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Registry resource address; defaults to backend config"
          }
        ],
        "responses": {
          "200": {
            "description": "Certificate record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CertificateViewRecord"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/marketplace/init": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_init",
        "summary": "Initialize MarketplaceStore under sender address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitMarketplaceRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/list": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_list",
        "summary": "List a certificate for sale",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ListCertificateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/cancel": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_cancel",
        "summary": "Cancel a listing (seller-only)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelListingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/request-buy": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_request_buy",
        "summary": "Request to buy a listed certificate",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestBuyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/accept-buy": {
      "post": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_accept_buy",
        "summary": "Seller accepts a pending buy request",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptBuyRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/marketplace/{market_addr}/stats": {
      "get": {
        "tags": [
          "Marketplace"
        ],
        "operationId": "market_stats",
        "summary": "Marketplace stats via view functions",
        "parameters": [
          {
            "name": "market_addr",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Stats",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MarketplaceStats"
                }
              }
            }
          }
        }
      }
    },
    "/audit/init": {
      "post": {
        "tags": [
          "Audit"
        ],
        "operationId": "audit_init",
        "summary": "Initialize audit config (enabled=true) under module owner address",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitAuditRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
    "/audit/log": {
      "post": {
        "tags": [
          "Audit"
        ],
        "operationId": "audit_log",
        "summary": "Emit an audit event (module event)",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuditLogRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Transaction submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TxResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "InitRegistryRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ]
      },
      "AddIssuerRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ],
        "type": "object",
        "properties": {
          "registry_addr": {
            "type": "string",
            "description": "Address where CertRegistry resource lives"
          },
          "issuer": {
            "type": "string",
            "description": "0x... issuer address"
          }
        },
        "required": [
          "registry_addr",
          "issuer"
        ]
      },
      "RemoveIssuerRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ],
        "type": "object",
        "properties": {
          "registry_addr": {
            "type": "string",
            "description": "Address where CertRegistry resource lives"
          },
          "issuer": {
            "type": "string",
            "description": "0x... issuer address"
          }
        },
        "required": [
          "registry_addr",
          "issuer"
        ]
      },
...
      "InitMarketplaceRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ]
      },
...
      "InitAuditRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/TxRequest"
          }
        ]
      },
...
    }
  }
}